% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_correlated_gaussian_target}
\alias{make_correlated_gaussian_target}
\title{Correlated Gaussian Ridge Target}
\usage{
make_correlated_gaussian_target()
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This two-dimensional Gaussian target stays in the package because it is the
cleanest first example of why covariance adaptation matters. If you want a
non-Gaussian ridge, use \code{make_correlated_t_target()}.
}
\examples{
ridge_target <- make_correlated_gaussian_target()
ridge_target$density(c(0, 0))

}
