% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_correlated_t_target}
\alias{make_correlated_t_target}
\title{Correlated Student-t Ridge Target}
\usage{
make_correlated_t_target(df = 4)
}
\arguments{
\item{df}{Degrees of freedom of the multivariate Student-t target. Smaller
values give heavier tails.}
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This target answers the obvious teaching question: the target does not have
to be Gaussian. The geometry is still a correlated ridge, but the heavier
tails make the target visibly non-Gaussian while keeping the mathematics and
contour plots easy to explain.
}
\examples{
target <- make_correlated_t_target()
target$density(c(0, 0))

}
