% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_dram_rescue_target}
\alias{make_dram_rescue_target}
\title{Delayed Rejection Teaching Target}
\usage{
make_dram_rescue_target()
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This two-dimensional Gaussian target is paired with a deliberately ambitious
first-stage proposal covariance. The setup is designed so that DRAM has a
genuine opportunity to rescue rejected large moves with a smaller second
stage.
}
\examples{
dram_target <- make_dram_rescue_target()
dram_target$recommended_delayed_rejection_scale

}
