% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_high_dimensional_gaussian_target}
\alias{make_high_dimensional_gaussian_target}
\title{Higher-Dimensional Gaussian Teaching Target}
\usage{
make_high_dimensional_gaussian_target(dimension = 8L, rho = 0.7)
}
\arguments{
\item{dimension}{Target dimension.}

\item{rho}{Correlation parameter used in the AR(1)-style covariance matrix.}
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This constructor creates a correlated Gaussian target in dimension \code{d}. It is
meant for demonstrating why manual proposal tuning becomes awkward in higher
dimensions and why RAM's acceptance-rate targeting can be attractive there.
}
\examples{
target <- make_high_dimensional_gaussian_target(dimension = 5)
target$dimension

}
