% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_multimodal_target}
\alias{make_multimodal_target}
\title{Multimodal Gaussian Mixture Target}
\usage{
make_multimodal_target(separation = 4, component_sd = 1)
}
\arguments{
\item{separation}{Distance between the two component means.}

\item{component_sd}{Standard deviation of each Gaussian mixture component.}
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This target is a simple symmetric Gaussian mixture. It is intentionally used
to show a limitation as well as a strength: adaptive random-walk methods can
improve local scale and covariance choices, but they do not automatically
solve multimodal trapping.
}
\examples{
target <- make_multimodal_target()
target$density(0)

}
