% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_small_scale_target}
\alias{make_small_scale_target}
\title{Poor Proposal Scale Target}
\usage{
make_small_scale_target()
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This one-dimensional Gaussian target is paired with an intentionally tiny
proposal variance. The result is usually a very high acceptance rate but very
slow exploration, which makes the example useful for discussing
autocorrelation.
}
\examples{
small_scale_target <- make_small_scale_target()
small_scale_target$recommended_proposal_covariance

}
