% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_standard_normal_target}
\alias{make_standard_normal_target}
\title{Standard Normal Teaching Target}
\usage{
make_standard_normal_target()
}
\value{
An object of class \code{"adaptive_mcmc_target"}.
}
\description{
This is the cleanest one-dimensional baseline in the package. It is useful
when you want to explain the mechanics of the samplers without also worrying
about correlation or multimodality.
}
\examples{
target <- make_standard_normal_target()
target$density(0)

}
