% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_acf_chain}
\alias{plot_acf_chain}
\title{Plot the Autocorrelation Function of One Chain}
\usage{
plot_acf_chain(run, dim = 1L, lag_max = 30L, burn_in = 0L)
}
\arguments{
\item{run}{A sampler result created by this package.}

\item{dim}{Coordinate to analyse.}

\item{lag_max}{Maximum lag displayed.}

\item{burn_in}{Number of initial iterations discarded.}
}
\value{
Invisibly returns the plotted autocorrelation values.
}
\description{
This single-chain ACF plot is useful when you want to explain one algorithm
at a time.
}
\examples{
comparison <- run_sampler_suite(make_standard_normal_target(), n_iter = 250, seed = 1)
plot_acf_chain(comparison$runs$basic, dim = 1, lag_max = 20)

}
