% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_jump_distance_histogram}
\alias{plot_jump_distance_histogram}
\title{Plot Jump Distance Histogram}
\usage{
plot_jump_distance_histogram(run, burn_in = 0L, breaks = "FD")
}
\arguments{
\item{run}{A sampler result created by this package.}

\item{burn_in}{Number of initial iterations to discard.}

\item{breaks}{Histogram break specification passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
Invisibly returns the jump distances used in the histogram.
}
\description{
Jump distances are especially useful in the poor-proposal-scale example.
}
\examples{
comparison <- run_sampler_suite(make_small_scale_target(), n_iter = 250, seed = 1)
plot_jump_distance_histogram(comparison$runs$basic)

}
