% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_target_contours}
\alias{plot_target_contours}
\title{Plot Contours of a Two-Dimensional Target}
\usage{
plot_target_contours(
  target,
  n_grid = 100L,
  levels = 10L,
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{target}{A two-dimensional target object created by this package.}

\item{n_grid}{Number of grid points in each coordinate direction.}

\item{levels}{Number of contour levels.}

\item{xlim}{Optional horizontal plot limits.}

\item{ylim}{Optional vertical plot limits.}

\item{add}{Should the contour plot be added to an existing figure?}

\item{...}{Additional arguments passed to \code{\link[graphics:contour]{graphics::contour()}}.}
}
\value{
Invisibly returns the grid used to draw the contours.
}
\description{
This helper evaluates the target density on a grid and draws contour lines that
can be overlaid with MCMC paths.
}
\examples{
target <- make_correlated_t_target()
plot_target_contours(target, n_grid = 60)

}
