% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Coordinate Traces}
\usage{
plot_trace(run, dims = NULL, burn_in = 0L)
}
\arguments{
\item{run}{A sampler result created by this package.}

\item{dims}{Coordinates to plot.}

\item{burn_in}{Number of initial iterations to omit.}
}
\value{
Invisibly returns the post-burn-in draws used in the plot.
}
\description{
Trace plots are a simple way to show sticking and slow exploration.
}
\examples{
comparison <- run_sampler_suite(make_small_scale_target(), n_iter = 250, seed = 1)
plot_trace(comparison$runs$basic)

}
