% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_chain}
\alias{summarize_chain}
\title{Summarize a Chain}
\usage{
summarize_chain(object, burn_in = 0L, ess_max_lag = NULL)
}
\arguments{
\item{object}{Either an \code{"adaptive_mcmc_run"} object or an
\code{"adaptive_mcmc_comparison"} object.}

\item{burn_in}{Number of initial iterations discarded before summarizing.}

\item{ess_max_lag}{Maximum lag used in the rough ESS calculation.}
}
\value{
A summary data frame.
}
\description{
This wrapper dispatches to \code{\link[=summarize_sampler_run]{summarize_sampler_run()}} or
\code{\link[=summarize_sampler_suite]{summarize_sampler_suite()}} depending on what you pass in.
}
\examples{
target <- make_standard_normal_target()
run <- run_metropolis(
  target$density,
  target$recommended_initial,
  200,
  target$recommended_proposal_covariance
)
summarize_chain(run)

}
